pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "python-dotenv>=1.0.0",
    "openai>=1.6.1",
    "websockets>=12.0",
    "pydantic>=2.5.3",
    "alembic>=1.13.1",
    "asyncio>=3.4.3",
    "python-multipart>=0.0.6",
    "ruff>=0.9.3",
    "sqlalchemy-stubs>=0.4",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
]

[tool.pytest]
pythonpath = "."
